update readme and polish code
